chore: fix project structure
