+++ src/components/AlbumsContainer.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import Album from './Album';
 
 function AlbumsContainer({ albums }) {
