+++ src/components/Artist.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-
 import { withRouter } from 'react-router';
 
 function Artist({ artist, history }) {
