+++ src/components/ArtistsContainer.jsx
