+++ src/components/PlaylistsContainer.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
-
 import Playlist from './Playlist';
 
 function PlaylistsContainer({ playlists, noSlide }) {
