+++ src/store/index.js
@@ -1,9 +1,9 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 
-import homeReducer from './reducers/home';
-import playerReducer from './reducers/player';
-import searchReducer from './reducers/search';
+import homeReducer from './home/reducer';
+import playerReducer from './player/reducer';
+import searchReducer from './search/reducer';
 
 const reducers = combineReducers({
   homeReducer,
